Slightly reduce circular imports
This improves the number of parallel jobs that can run as part of pytype
since it requires all imports (even those nested inside functions) to
be evaluated before proceeding to the next file.
