+++ pycheribuild/projects/cherios.py
@@ -30,8 +30,10 @@
 
 from .cmake_project import CMakeProject
 from .project import BuildType, ComputedDefaultValue, GitRepository
+from .run_qemu import LaunchQEMUBase, get_default_ssh_forwarding_port
 from .simple_project import BoolConfigOption, IntConfigOption
 from ..config.compilation_targets import CompilationTargets
+from ..utils import OSInfo
 
 
 class BuildCheriOS(CMakeProject):
@@ -39,9 +41,10 @@ class BuildCheriOS(CMakeProject):
     default_build_type = BuildType.DEBUG
     repository = GitRepository("https://github.com/CTSRD-CHERI/cherios.git", default_branch="master")
     _default_install_dir_fn = ComputedDefaultValue(
-        function=lambda config, p: config.output_root / ("cherios" +
-                                                         p.crosscompile_target.build_suffix(config, include_os=False)),
-        as_string="$OUTPUT_ROOT/cherios-{mips64,riscv64}")
+        function=lambda config, p: config.output_root
+        / ("cherios" + p.crosscompile_target.build_suffix(config, include_os=False)),
+        as_string="$OUTPUT_ROOT/cherios-{mips64,riscv64}",
+    )
     needs_sysroot = False
     supported_architectures = (CompilationTargets.CHERIOS_MIPS_PURECAP, CompilationTargets.CHERIOS_RISCV_PURECAP)
 
@@ -60,3 +63,55 @@ class BuildCheriOS(CMakeProject):
 
     def install(self, **kwargs):
         pass  # nothing to install yet
+
+
+class LaunchCheriOSQEMU(LaunchQEMUBase):
+    target = "run-cherios"
+    dependencies = ("qemu", "cherios")
+    supported_architectures = (CompilationTargets.CHERIOS_MIPS_PURECAP, CompilationTargets.CHERIOS_RISCV_PURECAP)
+    forward_ssh_port = False
+    qemu_user_networking = False
+    hide_options_from_help = True
+
+    @classmethod
+    def setup_config_options(cls, **kwargs):
+        super().setup_config_options(default_ssh_port=get_default_ssh_forwarding_port(40), **kwargs)
+
+    @property
+    def source_project(self):
+        return BuildCheriOS.get_instance(self, self.config)
+
+    def setup(self):
+        super().setup()
+        # FIXME: these should be config options
+        cherios = BuildCheriOS.get_instance(self, self.config)
+        self.current_kernel = cherios.build_dir / "boot/cherios.elf"
+        self.disk_image = self.config.output_root / "cherios-disk.img"
+        self._project_specific_options = ["-no-reboot", "-global", "virtio-mmio.force-legacy=false"]
+
+        if cherios.build_net:
+            self._after_disk_options.extend([
+                "-netdev",
+                "tap,id=tap0,ifname=cherios_tap,script=no,downscript=no",
+                "-device",
+                "virtio-net-device,netdev=tap0",
+            ])
+
+        if cherios.smp_cores > 1:
+            self._project_specific_options.append("-smp")
+            self._project_specific_options.append(str(cherios.smp_cores))
+
+        self.qemu_options.virtio_disk = True  # CheriOS needs virtio
+        self.qemu_options.force_virtio_blk_device = True
+        self.qemu_user_networking = False
+
+    def process(self):
+        if not self.disk_image.exists():
+            if self.query_yes_no("CheriOS disk image is missing. Would you like to create a zero-filled 1MB image?"):
+                size_flag = "bs=128m" if OSInfo.IS_MAC else "bs=128M"
+                self.run_cmd("dd", "if=/dev/zero", "of=" + str(self.disk_image), size_flag, "count=1")
+        super().process()
+
+    def get_riscv_bios_args(self) -> "list[str]":
+        # CheriOS bundles its kernel with its own bootloader
+        return ["-bios", "none"]
